content_factory_2026/v2/requirements.txt
openai>=1.30.0
anthropic>=0.25.0
groq>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
